+++ api/build.gradle.kts
@@ -11,7 +11,7 @@ dependencies {
     annotationProcessor("org.springframework.boot:spring-boot-configuration-processor")
     testImplementation("org.springframework.boot:spring-boot-starter-test")
 
-    implementation(project(":utilities"))
+    implementation(project(":commons"))
 }
